pid, id for UI, Network and Web module, checkbox type using bool3State
RAM:   [==        ]  17.6% (used 57748 bytes from 327680 bytes)
Flash: [========  ]  75.7% (used 1438593 bytes from 1900544 bytes)

index.js:
- initCap: add isLowerCase function (WIP)
- createHTML: check variable.value != null (bugfix)

newui/modules.js
- check variable.value != null (bugfix)

model.json
- rename clTbl to clients including columns
- rename ethOn, WifiOn, apOn to ethernet, wiFi, AP including children
- rename vlTbl to loops including columns

SysModModel
- showObsolete, don't use class variable
- callVarOnChange: checkbox pointer: use bool3State type

SysModNetwork
- rename ethOn, WifiOn, apOn to ethernet, wiFi, AP including children

SysModUI
- rename vlTbl to loops including columns
- initCheckbox use bool3State type
- callVarFun: use bool3State type

SysModWeb
- rename clTbl to clients including columns

+++ src/SysModules.cpp
@@ -50,8 +50,8 @@ void SysModules::setup() {
     default: return false;
   }});
 
-  //UINT16_MAX: no value set
-  ui->initCheckBox(tableVar, "success", UINT16_MAX, true, [this](JsonObject var, uint8_t rowNr, uint8_t funType) { switch (funType) { //varFun
+  //UINT8_MAX: tri state boolean: not true not false
+  ui->initCheckBox(tableVar, "success", UINT8_MAX, true, [this](JsonObject var, uint8_t rowNr, uint8_t funType) { switch (funType) { //varFun
     case onSetValue:
       for (size_t rowNr = 0; rowNr < modules.size(); rowNr++)
         mdl->setValue(var, modules[rowNr]->success, rowNr);
@@ -59,7 +59,8 @@ void SysModules::setup() {
     default: return false;
   }});
 
-  ui->initCheckBox(tableVar, "enabled", UINT16_MAX, false, [this](JsonObject var, uint8_t rowNr, uint8_t funType) { switch (funType) { //varFun not readonly! (tbd)
+  //UINT8_MAX: tri state boolean: not true not false
+  ui->initCheckBox(tableVar, "enabled", UINT8_MAX, false, [this](JsonObject var, uint8_t rowNr, uint8_t funType) { switch (funType) { //varFun not readonly! (tbd)
     case onSetValue:
       //never a rowNr as parameter, set all
       //execute only if var has not been set
